More warning fixes.
FossilOrigin-Name: 51a6a797e1d61c4361b87cdf3467bfd3ce161cc3

+++ src/test6.c
@@ -637,7 +637,7 @@ static int cfOpen(
 
       memset(pWrapper->zData, 0, pWrapper->nData);
       for(iOff=0; iOff<pWrapper->iSize; iOff += 512){
-        int nRead = pWrapper->iSize - iOff;
+        int nRead = pWrapper->iSize - (int)iOff;
         if( nRead>512 ) nRead = 512;
         if( isDb && iOff==PENDING_BYTE ) continue;
         rc = sqlite3OsRead(pReal, &pWrapper->zData[iOff], nRead, iOff);
